// modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/ModelBuilderContext/typesEnum.ts
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export enum TYPES {
	ADD_NEW_NODE_TO_FOLDER = 'ADD_NEW_NODE_TO_FOLDER',
	BULK_CHANGE_NODE_VIEW = 'BULK_CHANGE_NODE_VIEW',
	CHANGE_NODE_VIEW = 'CHANGE_NODE_VIEW',
	CREATE_MODEL_BUILDER_STRUCTURE = 'CREATE_MODEL_BUILDER_STRUCTURE',
	DELETE_FOLDER_NODE = 'DELETE_FOLDER_NODE',
	SET_ELEMENTS = 'SET_ELEMENTS',
	SET_SHOW_CHANGES_SAVED = 'SET_SHOW_CHANGES_SAVED',
	SET_SELECTED_EDGE = 'SET_SELECTED_EDGE',
	SET_SELECTED_NODE = 'SET_SELECTED_NODE',
	UPDATE_FOLDER_NODE = 'UPDATE_FOLDER_NODE',
}